move templates and size constants to a separate library (LifeModels)
