--- src/main/java/ru/life/constant/Size.java
@@ -1,16 +1,0 @@
-package ru.life.constant;
-
-public class Size {
-
-    public static int DOT_SIZE = 8;
-
-    public static final int SIZE_HEIGHT = 560;
-    public static final int SIZE_WIDTH = 880;
-
-    public static final int FRAME_SIZE_HEIGHT = 620;
-    public static final int FRAME_SIZE_WIDTH = 894;
-
-    public static final int DIALOG_FRAME_SIZE_HEIGHT = 600;
-    public static final int DIALOG_FRAME_SIZE_WIDTH = 300;
-
-}
